Initial restructure of groups section
