+++ src/components/Groups/CourseItem.tsx
@@ -2,15 +2,15 @@ import { Text } from 'office-ui-fabric-react';
 import { Card, ICardTokens } from "@uifabric/react-cards";
 import { initializeIcons } from '@uifabric/icons';
 import { FontWeights, ITextStyles, Persona } from '@fluentui/react';
-import { semibold } from '../services/fonts';
+import { semibold } from '../../services/fonts';
 import { useTheme } from '@fluentui/react-theme-provider';
 import { IContextualMenuProps, IIconProps } from '@fluentui/react';
 import { CommandButton } from '@fluentui/react/lib/Button';
 import { ActionButton } from '@fluentui/react/lib/Button';
-import { redirectToLink } from '../services/Utils';
-import { CourseDegree } from '../models/Models';
+import { redirectToLink } from '../../services/Utils';
+import { CourseDegree } from '../../models/Models';
 import Chip from '@material-ui/core/Chip';
-import LocalizationService from "../services/LocalizationService";
+import LocalizationService from "../../services/LocalizationService";
 
 initializeIcons();
 interface Props { data: CourseDegree };
