+++ src/views/ContentView.tsx
@@ -1,7 +1,6 @@
 import HomeView from './HomeView';
-import CoursesView from './CoursesView';
+import GroupsView from './GroupsView';
 import NewsView from './NewsView';
-import AdditionalGroupsView from './AdditionalGroupsView';
 import OrganizationView from './OrganizationView';
 import RulesView from './RulesView';
 import ServicesView from './ServicesView';
@@ -18,9 +17,8 @@ const ContentView = () => {
             <Route path="/home" render={() => wrap(<HomeView/>)} />
             <Route path="/rules" render={() => wrap(<RulesView/>)} />
             <Route path='/news' render={() => wrap(<NewsView/>)} />
-            <Route path='/courses' render={() => wrap(<CoursesView/>)} />
+            <Route path='/courses' render={() => wrap(<GroupsView/>)} />
             <Route path='/services' render={() => wrap(<ServicesView/>)} />
-            <Route path='/additional_groups' render={() => wrap(<AdditionalGroupsView/>)} />
             <Route path='/representatives' render={() => wrap(<RepresentativesView/>)} />
             <Route path='/organization' render={() => wrap(<OrganizationView/>)} />
         </Switch>
